Split off Event rendering from AlertPopup
