+++ core/src/com/unciv/ui/screens/worldscreen/RenderEvent.kt
@@ -1,0 +1,78 @@
+package com.unciv.ui.screens.worldscreen
+
+import com.badlogic.gdx.scenes.scene2d.ui.Table
+import com.badlogic.gdx.utils.Align
+import com.unciv.models.ruleset.Event
+import com.unciv.models.ruleset.EventChoice
+import com.unciv.models.ruleset.unique.StateForConditionals
+import com.unciv.ui.components.UncivTooltip.Companion.addTooltip
+import com.unciv.ui.components.extensions.addSeparator
+import com.unciv.ui.components.extensions.toTextButton
+import com.unciv.ui.components.input.KeyCharAndCode
+import com.unciv.ui.components.input.keyShortcuts
+import com.unciv.ui.components.input.onActivation
+import com.unciv.ui.components.widgets.WrappableLabel
+import com.unciv.ui.screens.civilopediascreen.FormattedLine
+import com.unciv.ui.screens.civilopediascreen.MarkupRenderer
+
+/** Renders an [Event] for [AlertPopup] or a floating tutorial task on [WorldScreen] */
+class RenderEvent(
+    event: Event,
+    val worldScreen: WorldScreen,
+    val onChoice: (EventChoice) -> Unit
+) : Table() {
+    private val gameInfo get() = worldScreen.gameInfo
+    private val stageWidth get() = worldScreen.stage.width
+
+    val isValid: Boolean
+
+    //todo check generated translations
+
+    init {
+        defaults().fillX().center().pad(5f)
+
+        val stateForConditionals = StateForConditionals(gameInfo.currentPlayerCiv)
+        val choices = event.getMatchingChoices(stateForConditionals)
+        isValid = choices != null
+        if (isValid) {
+            if (event.text.isNotEmpty()) {
+                add(WrappableLabel(event.text, stageWidth * 0.5f).apply {
+                    wrap = true
+                    setAlignment(Align.center)
+                    optimizePrefWidth()
+                }).row()
+            }
+            if (event.civilopediaText.isNotEmpty()) {
+                add(event.renderCivilopediaText(stageWidth * 0.5f, ::openCivilopedia)).row()
+            }
+
+            for (choice in choices!!) addChoice(choice)
+        }
+    }
+
+    private fun addChoice(choice: EventChoice) {
+        addSeparator()
+
+        val button = choice.text.toTextButton()
+        button.onActivation {
+            onChoice(choice)
+            choice.triggerChoice(gameInfo.currentPlayerCiv)
+        }
+        val key = KeyCharAndCode.parse(choice.keyShortcut)
+        if (key != KeyCharAndCode.UNKNOWN) {
+            button.keyShortcuts.add(key)
+            button.addTooltip(key)
+        }
+        add(button).row()
+
+        val lines = (
+            choice.civilopediaText.asSequence()
+                + choice.triggeredUniqueObjects.asSequence()
+                .filterNot { it.isHiddenToUsers() }
+                .map { FormattedLine(it) }
+            ).asIterable()
+        add(MarkupRenderer.render(lines, stageWidth * 0.5f, linkAction = ::openCivilopedia)).row()
+    }
+
+    private fun openCivilopedia(link: String) = worldScreen.openCivilopedia(link)
+}
